Add files via upload
Hi max, 
These are the programs for identifying the probability distribution of data.  At the end of the program, I haven't added the saving option because mostly i used it to verify the outcome of the program in the console and save the final output. if you have any questions please let me know.